Move *ExpectedType traits into traits.rs
